add aduan ops

+++ Controllers/Situgu/Ops/Cs.php
@@ -67,13 +67,22 @@ class Cs extends BaseController
             $row = [];
 
             $row[] = $no;
-            $action = '<div class="btn-group">
+            if ($list->status_ajuan == 2) {
+                $action = '<div class="btn-group">
+                        <button type="button" class="btn btn-primary dropdown-toggle" data-bs-toggle="dropdown" aria-expanded="false">Action <i class="mdi mdi-chevron-down"></i></button>
+                        <div class="dropdown-menu" style="">
+                            <a class="dropdown-item" href="javascript:actionDetail(\'' . $list->id . '\');"><i class="bx bxs-show font-size-16 align-middle"></i> &nbsp;Detail</a>
+                        </div>
+                    </div>';
+            } else {
+                $action = '<div class="btn-group">
                         <button type="button" class="btn btn-primary dropdown-toggle" data-bs-toggle="dropdown" aria-expanded="false">Action <i class="mdi mdi-chevron-down"></i></button>
                         <div class="dropdown-menu" style="">
                             <a class="dropdown-item" href="javascript:actionDetail(\'' . $list->id . '\');"><i class="bx bxs-show font-size-16 align-middle"></i> &nbsp;Detail</a>
                             <a class="dropdown-item" href="javascript:actionHapus(\'' . $list->id . '\');"><i class="bx bx-trash font-size-16 align-middle"></i> &nbsp;Hapus</a>
                         </div>
                     </div>';
+            }
             // $action = '<a href="javascript:actionDetail(\'' . $list->id . '\', \'' . str_replace('&#039;', "`", str_replace("'", "`", $list->nama)) . '\');"><button type="button" class="btn btn-primary btn-sm btn-rounded waves-effect waves-light mr-2 mb-1">
             //     <i class="bx bxs-show font-size-16 align-middle"></i></button>
             //     </a>
